Add parsed kernel information to UnameResult
